requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
groq==0.4.2
openai-whisper==20231117
langdetect==1.0.9 
